Added elysia-autoroutes as a dependency and changed the code to work with the file system routing. Also includes updated versios of elysia and @elysiajs/html
