+++ src/components/todo-item.tsx
@@ -1,0 +1,25 @@
+import * as elements from "typed-html";
+import { Todo } from "../db/schema";
+
+export function TodoItem({ content, completed, id }: Todo) {
+  return (
+    <div class="flex flex-row space-x-3">
+      <p>{content}</p>
+      <input
+        type="checkbox"
+        checked={completed}
+        hx-post={`/todos/toggle/${id}`}
+        hx-swap="outerHTML"
+        hx-target="closest div"
+      />
+      <button
+        class="text-red-500"
+        hx-delete={`/todos/${id}`}
+        hx-swap="outerHTML"
+        hx-target="closest div"
+      >
+        X
+      </button>
+    </div>
+  );
+}
