+++ src/html.ts
@@ -1,0 +1,17 @@
+import * as elements from "typed-html";
+
+export const BaseHtml = ({ children }: elements.Children) => `
+<!DOCTYPE html>
+<html lang="en">
+
+<head>
+  <meta charset="UTF-8">
+  <meta name="viewport" content="width=device-width, initial-scale=1.0">
+  <title>THE BETH STACK</title>
+  <script src="https://unpkg.com/htmx.org@1.9.3"></script>
+  <script src="https://unpkg.com/hyperscript.org@0.9.9"></script>
+  <link href="/styles.css" rel="stylesheet">
+</head>
+
+${children}
+`;
