--- src/index.tsx
@@ -1,164 +1,0 @@
-import { Elysia, t } from "elysia";
-import { html } from "@elysiajs/html";
-import * as elements from "typed-html";
-import { db } from "./db";
-import { Todo, todos } from "./db/schema";
-import { eq } from "drizzle-orm";
-
-const app = new Elysia()
-  .use(html())
-  .get("/", ({ html }) =>
-    html(
-      <BaseHtml>
-        <body
-          class="flex w-full h-screen justify-center items-center"
-          hx-get="/todos"
-          hx-swap="innerHTML"
-          hx-trigger="load"
-        />
-      </BaseHtml>
-    )
-  )
-  .get("/todos", async () => {
-    const data = await db.select().from(todos).all();
-    return <TodoList todos={data} />;
-  })
-  .post(
-    "/todos/toggle/:id",
-    async ({ params }) => {
-      const oldTodo = await db
-        .select()
-        .from(todos)
-        .where(eq(todos.id, params.id))
-        .get();
-      const newTodo = await db
-        .update(todos)
-        .set({ completed: !oldTodo.completed })
-        .where(eq(todos.id, params.id))
-        .returning()
-        .get();
-      return <TodoItem {...newTodo} />;
-    },
-    {
-      params: t.Object({
-        id: t.Numeric(),
-      }),
-    }
-  )
-  .delete(
-    "/todos/:id",
-    async ({ params }) => {
-      await db.delete(todos).where(eq(todos.id, params.id)).run();
-    },
-    {
-      params: t.Object({
-        id: t.Numeric(),
-      }),
-    }
-  )
-  .post(
-    "/todos",
-    async ({ body }) => {
-      const content = {
-        beth: "Learn the BETH stack",
-        vim: "Learn vim",
-        like: "Like the video",
-        sub: "Subscribe to Ethan",
-      };
-
-      const newTodo = await db
-        .insert(todos)
-        .values({ content: content[body.content] })
-        .returning()
-        .get();
-      return <TodoItem {...newTodo} />;
-    },
-    {
-      body: t.Object({
-        content: t.Union([
-          t.Literal("beth"),
-          t.Literal("vim"),
-          t.Literal("like"),
-          t.Literal("sub"),
-        ]),
-      }),
-    }
-  )
-  .get("/styles.css", () => Bun.file("./tailwind-gen/styles.css"))
-  .listen(3000);
-
-console.log(
-  `🦊 Elysia is running at http://${app.server?.hostname}:${app.server?.port}`
-);
-
-const BaseHtml = ({ children }: elements.Children) => `
-<!DOCTYPE html>
-<html lang="en">
-
-<head>
-  <meta charset="UTF-8">
-  <meta name="viewport" content="width=device-width, initial-scale=1.0">
-  <title>THE BETH STACK</title>
-  <script src="https://unpkg.com/htmx.org@1.9.3"></script>
-  <script src="https://unpkg.com/hyperscript.org@0.9.9"></script>
-  <link href="/styles.css" rel="stylesheet">
-</head>
-
-${children}
-`;
-
-function TodoItem({ content, completed, id }: Todo) {
-  return (
-    <div class="flex flex-row space-x-3">
-      <p>{content}</p>
-      <input
-        type="checkbox"
-        checked={completed}
-        hx-post={`/todos/toggle/${id}`}
-        hx-swap="outerHTML"
-        hx-target="closest div"
-      />
-      <button
-        class="text-red-500"
-        hx-delete={`/todos/${id}`}
-        hx-swap="outerHTML"
-        hx-target="closest div"
-      >
-        X
-      </button>
-    </div>
-  );
-}
-
-function TodoList({ todos }: { todos: Todo[] }) {
-  return (
-    <div>
-      {todos.map((todo) => (
-        <TodoItem {...todo} />
-      ))}
-      <TodoForm />
-    </div>
-  );
-}
-
-function TodoForm() {
-  return (
-    <form
-      class="flex flex-row space-x-3"
-      hx-post="/todos"
-      hx-swap="beforebegin"
-      _="on submit target.reset()"
-    >
-      <select name="content" class="border border-black">
-        <option value="" disabled="true" selected="true">
-          Select a Todo
-        </option>
-        <option value="beth">Learn the BETH stack</option>
-        <option value="vim">Learn vim</option>
-        <option value="like">Like the video</option>
-        <option value="sub">Subscribe to Ethan</option>
-      </select>
-      <button type="submit">Add</button>
-    </form>
-  );
-}
